+++ packages/terra-core-docs/src/terra-dev-site/test/heading/HeadingItalics.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Heading from 'terra-heading';
+
+const HeadingItalics = () => (
+  <div>
+    <Heading id="heading-italic" level={1} isItalic>Italics</Heading>
+  </div>
+);
+
+export default HeadingItalics;