+++ packages/terra-core-docs/src/terra-dev-site/test/heading/HeadingLevels.test.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Heading from 'terra-heading';
+
+const HeadingLevels = () => (
+  <div>
+    <Heading id="heading-level-1" level={1}>Level 1</Heading>
+    <Heading id="heading-level-2" level={2}>Level 2</Heading>
+    <Heading id="heading-level-3" level={3}>Level 3</Heading>
+    <Heading id="heading-level-4" level={4}>Level 4</Heading>
+    <Heading id="heading-level-5" level={5}>Level 5</Heading>
+    <Heading id="heading-level-6" level={6}>Level 6</Heading>
+  </div>
+);
+
+export default HeadingLevels;