+++ packages/terra-core-docs/src/terra-dev-site/test/heading/HeadingVisuallyHidden.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Heading from 'terra-heading';
+
+const HeadingVisuallyHidden = () => (
+  <div>
+    <span>Heading below is visually hidden but available to screen readers</span>
+    <Heading id="heading-visually-hidden" level={1} isVisuallyHidden>Visually Hidden</Heading>
+  </div>
+);
+
+export default HeadingVisuallyHidden;